Revert "Move LayersScraper"
This reverts commit 5848f90ec0de806957da2ec1d6a05fec37c0a9dd.
